pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-test-framework"
version = "0.2.0"
description = "Production-ready async API testing framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.28.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "allure-pytest>=2.13.5",
  "faker>=30.0.0",
  "jsonschema>=4.23.0",
  "pytest-xdist>=3.6.0",
  "eval_type_backport>=0.2.0",
  "schemathesis>=3.37.0",
  "pytest-cov>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=4.0.0",
  "ruff>=0.7.0",
  "black>=24.10.0",
  "isort>=5.13.0",
  "mypy>=1.13.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = "-ra -q"
testpaths = ["tests"]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
log_cli_date_format = "%H:%M:%S"

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
plugins = []
